pyproject.toml
[tool.poetry]
name = "sjtop"
version = "0.1.2"
homepage = "https://github.com/yvictor/sjtop"
description = "Terminal User Interface for Shioaji"
authors = ["Yvictor <yvictor3141@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.scripts]
sjtop = 'sjtop.__main__:main'

[tool.poetry.dependencies]
python = "^3.7"
textual = "^0.1.12"
shioaji = ">=0.3.3.dev3"
pandas = ">=0.24.1"
pytest-freezegun = "^0.4.2"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
black = "^21.9b0"
pytest-cov = "^2.12.1"
pytest-mock = "^3.6.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
